# [quant] Speed up per-channel min-max observer