--- a/torch/quantization/observer.py
+++ b/torch/quantization/observer.py
@@ -144,21 +144,58 @@ class _ObserverBase(ObserverBase):
             )
             return torch.tensor([1.0]), torch.tensor([0])
 
-        for i in range(len(min_vals)):
-            assert (
-                min_vals[i] <= max_vals[i]
-            ), "min {} should be less than max {}".format(min_vals[i], max_vals[i])
+        diff = min_vals <= max_vals
+        for i in range(len(diff)):
+            assert(diff[i], "min should be less than max for index {}".format(i))
 
         scales = torch.empty(min_vals.size(), dtype=torch.float32)
         zero_points = torch.empty(min_vals.size(), dtype=torch.int64)
+        
+        if self.dtype == torch.qint8:
+            if self.reduce_range:
+                qmin, qmax = -64, 63
+            else:
+                qmin, qmax = -128, 127
+        else:
+            if self.reduce_range:
+                qmin, qmax = 0, 127
+            else:
+                qmin, qmax = 0, 255
 
+        max_vals, min_vals = max_vals.to(dtype=torch.float), min_vals.to(dtype=torch.float)
+
+        min_vals = torch.min(min_vals, torch.tensor([0.], device=min_vals.device))
+        max_vals = torch.max(max_vals, torch.tensor([0.], device=max_vals.device))
+        if torch.equal(max_vals, min_vals):
+            scales.fill_(1.0)
+            zero_points.fill_(0)
+        else:
+            if self.qscheme == torch.per_tensor_symmetric or self.qscheme == torch.per_channel_symmetric:
+                max_vals = torch.max(-min_vals, max_vals)
+                scales = max_vals / ((qmax - qmin) / 2)
+                scales = torch.max(scales, torch.tensor([self.eps], device=scales.device))
+                if self.dtype == torch.qint8:
+                    zp = 0
+                else:
+                    zp = 128
+                zero_points.fill_(zp)
+            else:
+                scales = (max_vals - min_vals) / float(qmax - qmin)
+                scales = torch.max(scales, torch.tensor([self.eps], device=scales.device))
+                zero_points = qmin - torch.round(min_vals / scales)
+                zero_points = torch.max(zero_points, torch.tensor([qmin], dtype=zero_points.dtype, device=zero_points.device))
+                zero_points = torch.min(zero_points, torch.tensor([qmax], dtype=zero_points.dtype, device=zero_points.device))
+                zero_points = zero_points.to(dtype=torch.int64)
+        scales.to(dtype=torch.float)
+        
+        '''
         for i in range(len(scales)):
             qparam = self._calculate_qparams(
                 min_vals[i], max_vals[i]
             )
             scales[i] = float(qparam[0])
             zero_points[i] = int(qparam[1])
-
+        '''
         return scales, zero_points
 
     @torch.jit.export