Stack from [ghstack](https://github.com/ezyang/ghstack):
* **#34118 [quant] Speed up per-channel min-max observer**

Summary: Previously calc_per_channel_qparams was using for loops and python primitives, which called `item` many times causing slowdown during training.
    These changes uses torch primitives on the tensor to speed up the operation over 60x

    Perf results on MobileNetV2 during training using autograd profiler

    FP32 forward call -
    Self CPU time total: 47.222ms
    CUDA time total: 124.001ms

    before change
    FakeQuant Model -
    Self CPU time total: 19.107s
    CUDA time total: 27.177s

    after change
    FakeQuant Model -
    Self CPU time total: 404.667ms
    CUDA time total: 446.344ms

Test Plan:
    python test/test_quantization.py


Reviewers:

Subscribers:

Tasks:

Tags:

Differential Revision: [D20287841](https://our.internmc.facebook.com/intern/diff/D20287841)